posts-grpc/requirements.txt
asyncpg
grpcio
grpcio-tools
sqlalchemy
